Fix issue with version table plugin outside of local checkout (#1301)
This commit fixes an issue with the version table plugin when running
a docs build outside a local checkout of the Qiskit/qiskit repository.
Previously in #1174 we switched to doing a fresh clone of the repository
from github to enable using the plugin outside of docs builds from
inside the local git repository (which is important for the
qiskit-translations repo which does not have a full clone). However,
that pr neglected to update the follow git calls used to get the
element versions (and as recently introduced in #1300 the release date)
to use that local checkout. So builds outside of the local repo have
been getting a full tag list but then trying to get the versions from
the local working directory's repository instead of the fresh clone.
This commit fixes this by making sure we use the git cloned version of
the qiskit meta-repository instead of the local working directory.