Remove pins on application module optional extras (#1315)
This commit removes the pinning on the optional extras for the
application packages. We shouldn't be doing this as these are downstream
dependencies of qiskit and we only provide these targets as an
installation convienence. As the packages are separate pinning things
just make this more complicated for qiskit to release (and leads to
extra package versions). This was mostly an artifact of the release
automation, but the behavior there is being updated in
Qiskit/qiskit-bot#14 to not bump anything around optional extras.

Depends-on: Qiskit/qiskit-bot#14