Add stable 0.19 docs to version list
It was requested that we have the 0.19 docs hosted on the documentation
page even though we only started doing stable doc builds at the 0.24
release. After a little trouble with dependency versions (things have
changed a bit in the past 14 months) I manually checked out, built, and
uploaded the 0.19 docs which are now hosted at:
https://qiskit.org/documentation/stable/0.19/ But, to make this version
discoverable we need to include 0.19 in the version list. This commit
manually prepends 0.19 to the version_list as it was a one off so it's
discoverable for users interested in reading the docs for the historical
version.