Revert "Introducing new qiskit modules naming (#1224)"
PR #1224 attempted to leave bread crumbs in the documentation to get
people used to the new package organization and split between Qiskit and
downstream packages (applications, experiments, etc). However, in
practice this has been causing more confusion for users than helping.
This primarily stemming from the use of the word "Experiments" to
describe ignis instead of pointing to qiskit-experiments. This reverts
that PR as the start of reoganizing things to be a bit clearer.

This reverts commit 3c4e44a2866e642d121a8298a62e814675cf0635.