Reorganize docs (#1323)
* Revert "Introducing new qiskit modules naming (#1224)"

PR #1224 attempted to leave bread crumbs in the documentation to get
people used to the new package organization and split between Qiskit and
downstream packages (applications, experiments, etc). However, in
practice this has been causing more confusion for users than helping.
This primarily stemming from the use of the word "Experiments" to
describe ignis instead of pointing to qiskit-experiments. This reverts
that PR as the start of reoganizing things to be a bit clearer.

This reverts commit 3c4e44a2866e642d121a8298a62e814675cf0635.

* Add details to home pages about downstream projects

* Add experiments to top bar and reformat home page

This commit adds experiments to the top bar where it was previously
missing and adjusts the formatting for the new home page sections on
applications and experiments.

* Updates from code review

* Improve description for application modules

* Update README to be more clear about current project statuses

* Adjust application order on index.rst

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Apply suggestions from code review

Co-authored-by: Ali Javadi-Abhari <ajavadia@users.noreply.github.com>