Add dynamics to other-builds.txt list (#1349)
The other-builds.txt list is used to maintain a list of directories to
not overwrite when we sync new docs builds to COS. These are other doc
builds hosted in the same namespace in the hosted bucket. With the
pending release of the qiskit-dynamics project it will be uploading it's
documentation to 'documentation/dynamics' and we need to add this to the
exclude list so that future docs updates for the root documentation do
not accidently delete the dynamics documentation. This commit adds
'dynamics/**'  to the other-builds.txt list for this reason.