Pin matplotlib in CI
With the recent release of matplotlib 3.5.0 bloch sphere visualizations
are failing in docs builds. This has been mitigated in terra's main
branch but hasn't been included in a release (but will be fixed in the
upcoming terra 0.19.0 release). Until that is released this commit pins
things in CI to avoid using the new version to unblock ci.