Revise wording in aqua removal release note prelude (#1371)
Thse section in the prelude for the 0.33.0 release contained a number of
copy and paste errors (as it was mostly copied from the deprecation note
in 0.25.0). This commit corrects those to make it clear that aqua has
been removed and you only need to uninstall aqua manually when upgrading
from a prior version.