Expand deprecation policy
This commit expands the deprecation policy to be a bit stricter and
clarify some edge cases that have arisen in recent releases. It adds
two sub-points to the previous deprecation policy:

1. A feature can't be deprecated until the alternative has been present
   in a released version.
2. A deprecated feature can't be removed until it's been deprecated in
   more than 1 release.

These two additional subpoint should clarify some ambiguity in the
current policy especially as we've had feedback on the current approach
to deprecation and typically have > 3 months between releases.