Set IFS only locally for read
Previously IFS was set for the whole bash setting, which later affected
the word splitting when the secrets were read from the variables and
split into the openssl command.

This also tidies up the array reading - previously, the `VERSION`
variable was an array (so `echo $VERSION` would be equivalent to
`echo ${VERSION[0]}`), and it was passed to `cut`, which isn't
necessary, since it was already split.