Update building from source instructions
With Qiskit/qiskit-terra#7658 and Qiskit/qiskit-terra#7702 not far
behind the requiremetns for building terra from source will be changed.
A C++ compiler is no longer required and instead a rust compiler is
needed. This commit updates the instructions on building from source and
also removes so old out of date notes from the document at the same
time.