Complete rewrite of Getting Started guide
This PR is the culmination of the documentation / on-boarding working
session from the Qiskit Dev conference in Tarrytown, April 2022.

The flagship changes are:

- the contributing guide is completely new, and targetted specifically
  at onboarding new external contributors, and helping them find issues
  to work on.  This includes a flowchart to direct people to
  repositories all across the Qiskit organisation, made by Abby and
  Anne.

- the old contributing guide is split up into separate pages, such as
  "building from source" and a "maintainer guide", so the information
  isn't quite so dense, and can be dipped into more easily.

- the deprecation policy is reorganised to make the processes more
  clear, and then later providing more details on how certain types of
  deprecation can be done.

More detail:

This makes the Getting Started guide the single (only) source of truth
for installing the metapackage components, moving the correct
information from the contributing guide to the getting started pages,
which were previously out-of-date.

This also splits the installation components of the getting started
pages into separate pages rather than separate tabs, because the
`..  tabbed::` directive didn't play nicely with Sphinx headings and
references, which caused all links to the source-installation sections
to be broken.

This also moves the deprecation policy into a separate file, and
rewrites the text to be shorter and more split it.  Also add sections on
the difficulty of changing behaviours and how to potentially avoid those
situations, and how to test the deprecated functionality correctly.

The actual policy should be unchanged by this commit.

Co-authored-by: Luciano Bello <bel@zurich.ibm.com>
Co-authored-by: Abby Mitchell <23662430+javabster@users.noreply.github.com>