Write separate deprecation guide
This moves the deprecation policy into a separate file, and rewrites the
text to be shorter and more split it.  Also add sections on the
difficulty of changing behaviours and how to potentially avoid those
situations, and how to test the deprecated functionality correctly.

The actual policy should be unchanged by this commit.