Use Terra's FakeBackend in benchmarks
Originally the benchmark suite created its own fake backends to avoid
noise if Terra were to add new snapshots.  The current old backends
use `BaseBackend` and `BaseJob`, which is the legacy interface that has
just been removed from mainline Terra, causing the benchmark suite to
fail to import.

The three fake backends in use are all either decommissioned machines
that will have no further updates, or a dummy machine not based on real
hardware (and so should not need to be updated).  Since it's unlikely
that noise will come into the benchmarks from Terra now, it's suitable
just to use the Terra-provided ones.