Change deprecated sphinx-panels to sphinx-design
`sphinx-panels` is unmaintained, and suggests `sphinx-design` as a
largely drop-in replacement.  For our uses (only within Terra), we need
nothing more than to just change the dependency.

`sphinx-panels` was previously pinning us to Sphinx < 5, transitively.
With the new version of Sphinx, setting `language = None` is deprecated;
English is no longer implied by `None`, we have to set it (since, you
know, there are other languages in the world...).