Separate stable version list from language menu (#1579)
* Separate stable version list from language menu

The language menu is using a widget/theme element rst-versions which in
the original rtd theme (which is the grandparent of the qiskit sphinx
theme) is a pop up/dropdown menu that has a bunch of extra information
in it that can be used to list different versions of the documentation
as well as download links. In #1261 that functionality was used to add a
list of stable versions to the sidebar since it was the logical place
for it. However, this has apparently been a source of confusion for most
users as unless they have prior knowledge about the previous version
links being on the same menu as the language selection and the
availability of old versions of the documentation goes unnoticed. This
commit addresses this by splitting the previous versions list out of the
language selector drop down and just adding a list to the bottom of the
side bar. We can change the formatting and style of this as needed in
the future, but at least for right now it gives us a clear location for
links to old version of the documenation.

Fixes #1266

* Update docs/_templates/stable_versions.html

Co-authored-by: Luciano Bello <bel@zurich.ibm.com>

Co-authored-by: Luciano Bello <bel@zurich.ibm.com>