Add rustworkx to other docs build list (#1583)
The retworkx project was recently renamed rustworkx (see
Qiskit/rustworkx#644) however the documentation is still being published
to the retworkx url. The project will soon start uploading builds to
qiskit.org/documentation/rustworkx in order to prepare for that this
commit adds the path to the exlude list so syncs of the main qiskit
documentation do not accidently delete the rustworkx documentation.