Redirect aer docs to standalone aer page
As part of the upcoming aer 0.11.0 release the project is moving to a
separate package/namespace and will no longer be part of `qiskit.*`.
This means the old documentation pages that existed for aer will no
longer be built because autodoc for the path will no longer be from
qiskit.providers.aer. To address this in the next metapackage release
this commit adds redirects from the existing aer documentation pages to
the new aer documentation.

Related to #1508