Switch to redirect in docs tree
To ease the migration of docs around the aer release this commit shifts
the redirect from the old import path to the new import path in the same
docs tree. This is needed for two reasons, while aer is still part of the
metapackage we should document it as such, and secondly the current state
of the Aer documentation isn't really for standalone hosting the docs
aren't really sufficient on their own yet. So until both situations are
resolved we should keep aer as part of the metapackage docs.