Add qiskit-aer documentation to other builds list (#1595)
This commit adds the qiskit aer documentation to the other builds path.
This will let qiskit-aer manage publishing its own documentation to
https://qiskit.org/documentation/aer
Starting in the pending 0.11.0 qiskit-aer release it has moved to its
own python package/namespace qiskit_aer which will complicate the
publishing the documentation under the same build as the qiskit main
docs. So aer is moving to publishing it's own doc page in
https://github.com/Qiskit/qiskit-aer/pull/1589