Fix deploying new docs to old stable tag (#1598)
* Fix deploying new docs to old stable tag

Previously, we relied on inferring the tag from `git describe`.  This
has been somewhat unreliable in the past, seemingly not recognising the
newest tag when triggered.  It is possible that the VM's git checkout
is initialised before the tag is synced into the main repo, which might
have caused this.

Instead, we extract the tag from the environment variable GitHub Actions
sets to pass to the deployment script directly.  This is guaranteed to
give the correct tag.

* Force use of bash