Redirect aer docs to new import path (#1596)
* Redirect aer docs to standalone aer page

As part of the upcoming aer 0.11.0 release the project is moving to a
separate package/namespace and will no longer be part of `qiskit.*`.
This means the old documentation pages that existed for aer will no
longer be built because autodoc for the path will no longer be from
qiskit.providers.aer. To address this in the next metapackage release
this commit adds redirects from the existing aer documentation pages to
the new aer documentation.

Related to #1508

* Switch to redirect in docs tree

To ease the migration of docs around the aer release this commit shifts
the redirect from the old import path to the new import path in the same
docs tree. This is needed for two reasons, while aer is still part of the
metapackage we should document it as such, and secondly the current state
of the Aer documentation isn't really for standalone hosting the docs
aren't really sufficient on their own yet. So until both situations are
resolved we should keep aer as part of the metapackage docs.