Fix redirect path for qiskit-aer modules
In #1596 we added redirects from the old qiskit aer module paths to the
new `qiskit_aer` path. However, that PR added the redirects relative to
the root of the docs build and the sphinx-reredirects module does the
redirects relative to the dir of the source path. This is causing the
destination of the redirects to go to the wrong URL resulting in a 404.
This commit updates the redirect configuration to ensure that the
destination path is correct by treating the destination path relative to
the source path which fixes the broken redirects.