Add s390x and ppc64le as tier 3 supported platforms (#1606)
With the recent release of qiskit-aer 0.11.0 all the qiskit packages
with compiled extensions now are publishing precompiled binaries for
s390x and ppc64le. This commit updates the supported platform
documentation to indicate that we support these platforms at tier 3,
meaning we build and publish binaries at release time but do not have CI
testing or infrastructure to test the binares post build.