Reduce number of permutations for assembler benchmarks
When running the full benchmark suite the assembler benchmarks take a
disproportionately large amount of time to run, especially given the
relatively low current importance of the assemble() function. This was
primarily due to the number of permutations of the benchmark we ran. At
the time all those permutations were added to verify the scaling of the
function because it was a potential bottleneck, but this is no longer a
concern. This commit drops the parameter sweeps down so that we only run
2 combinations for assemble, basically a fixed sized circuit either by
itself or as a batch of 100. This should drastically decrease the
runtime cost of these benchmarks but still give us some data for the
function moving forward.