Fix install from source docs (#1625)
* Fix install from source docs

In #1439 we updated the install from source documentation to reflect
recent (as in the last year) changes to the build process and ensure the
documentation was correct for installing from source. However, in #1507
where some of the documentation was refactored these changes were
inadvertadntly delete. This commit restores the correct source
installation instructions from #1439 so that the hosted docs correctly
document how to build and install qiskit from source.

Fixes #1624

* Update docs/getting_started.rst

* Remove unnecessary pip install from aer instructions