Use -d flag on sphinx-build to move .doctrees dir outside build
This commit updates our default sphinx-build invocation to include the
-d flag to change the default output directory for the .doctrees dir
(see: https://www.sphinx-doc.org/en/master/man/sphinx-build.html#cmdoption-sphinx-build-d
for details on the flag). The .doctrees dir is a local cache of pickle
files for parsed source files and doesn't need to be included in the
build. By moving the default location for this outside our build dir
this will remove a lot of overhead from our uploaded docs as this cache
can end up being quite large.