Add deprecate to qiskit-ibmq-provider entry on the sidebar (#1661)
* Fix toctree issues with documentation

In #1631 the docs were updated to use a new version of the
qiskit_sphinx_theme and enable some new features it included. However,
in doing this a few issues were introduced to the documentation. The
biggest of which was the removal of the aer, circuit library, and the
ibmq provider from the table of contents tree. Looking at the review
comments on #1631 this was a consious choice to rely on hard coded
navigation in the theme (i.e. the top bar). However, doing this breaks
navigation in the sphinx documentation because it relies on the toctree
to build the pathing for navigation from leaf pages in the toctree.
For example, if you want to find the QuantumVolume class documentation
currently you must search to find it because you can't get to it from
any other page. Then when you find you're not able to get back to the
parent page for the circuit library because it's an orphaned page
outside the toctree. While aer and the ibmq provider get around this to
a certain degree because you can get the per project index via the top
bar (which I personally find unintuitive) navigation within that
namespace is broken because they're not reachable from the root index.
This commit adds back the missing references back to the root table of
contents so navigation works again.

The other changes are relatively minor as it just fixes up the FAQ page
which I mostly didn't realize existed prior to looking at the docs just
now. Primarily it cleans up some irrelevant questions and also uses
proper sphinx references to point to the install guide when needed.

* Update docs/faq.rst

Co-authored-by: Kevin Hartman <kevin@hart.mn>

* Update docs/index.rst

* Fix more missing pages

* Split out API references into separate toctree

* Fix typo in index.rst

Co-authored-by: Junye Huang <h.jun.ye@gmail.com>

* add latest agreed sidebar

* add back missing overview and qc_intro page

* make diff cleaner

* remove unnecessary maxdepth

* add explicit text to Aer and ibmq provider api links

* add deprecation to IBMQ provider api entry on the sidebar

* change provider names

* Update docs/index.rst

Co-authored-by: Luciano Bello <bel@zurich.ibm.com>

---------

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
Co-authored-by: Kevin Hartman <kevin@hart.mn>
Co-authored-by: Luciano Bello <bel@zurich.ibm.com>