Fix `tox -e docs` for macOS (#1694)
<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary

`tox -e docs` now works on my M1.

### Details and comments

We now say to install `.[all]` in two places: before installing the package itself when just installing `deps`, and then again when installing the package itself. Tox uses the same venv for both times, so that essentially results in a no-op in the second stage. This is similar to https://github.com/Qiskit/qiskit-terra/pull/9758.