Use templates and source_images folders from Terra (#1748)
<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

This makes progress towards https://github.com/Qiskit/qiskit-metapackage/issues/1722. 

We are not ready to yet move actual docs pages like `maintainers_guide.rst` to Terra because it requires that those docs are in a released version of Terra, such as 0.24.1.

But, Terra has already had duplicate copies of `_templates` and `source_images`. So, it's safe for us to de-duplicate those folders.

This PR is prework for future de-duplication.

--

To simplify this PR, I removed the `DOCS_FROM_MASTER` feature, which was not used in CI or anywhere.

I also tweaked the functionality of docs currently existing. We now will delete them at the end of the run. And we don't clone tutorials if they already existed.