Add intersphinx extension for linking to other orgs (#1769)
<!--
⚠️ If you do not respect this template, your pull request will be
closed.
⚠️ Your pull request title should be short detailed and understandable
for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary

Adding intersphinx extension to allow for linking to other orgs in docs


### Details and comments

I added links for `matplotlib.figure.Figure` return types to the
visualization docs in https://github.com/Qiskit/qiskit-terra/pull/10389
, and @Eric-Arellano requested in
https://github.com/Qiskit/qiskit-terra/pull/10389#pullrequestreview-1518994619
to make that change here as well.

It didn't appear that the intersphinx extension was being used in this
repo (I didn't find it in `docs/conf.py`), so I added the extension and
all the mappings from qiskit-terra. Let me know if that's not the
correct approach to take here.

---------

Signed-off-by: Paul S. Schweigert <paul@paulschweigert.com>