Fix add_module_names usage in conf.py
As part of the the recently merged #1784 we tried to update the
add_module_names flag to be true to correspond to some organizational
changes made to the docs in the release. However, a typo in that PR
resulted in not actually setting the flag correctly. It also neglected
that it was already set in the configuration file above to False. This
commit removes the duplicate typo entry and just updates the existing
entry to be the new setting.