Fix add_module_names usage in conf.py (#1786)
<!--
⚠️ If you do not respect this template, your pull request will be
closed.
⚠️ Your pull request title should be short detailed and understandable
for all.
⚠️ Also, please add it in the CHANGELOG file under Unreleased section.
⚠️ If your pull request fixes an open issue, please link to the issue.

✅ I have added the tests to cover my changes.
✅ I have updated the documentation accordingly.
✅ I have read the CONTRIBUTING document.
-->

### Summary

As part of the the recently merged #1784 we tried to update the
add_module_names flag to be true to correspond to some organizational
changes made to the docs in the release. However, a typo in that PR
resulted in not actually setting the flag correctly. It also neglected
that it was already set in the configuration file above to False. This
commit removes the duplicate typo entry and just updates the existing
entry to be the new setting.

### Details and comments