Remove metapackage-specific parts of documentation
This removes many sections of the documentation that refer to the
metapackage as a whole, the elements of the metapackage, or refer to
Terra as Terra.  This is intended to be in line with the new branding
decisions around the Qiskit project.

With the current community management structure still a little up in the
air, we are (at least temporarily) removing the repository-decision
flowchart from the documentation, since this documentation will
primarily be for Qiskit Terra.  The contributor flowchart, or something
similar, may end up on the landing pages of the Ecosystem documentation,
or another community hub.